\name{fwInfo}


\alias{fwInfo}

\title{Describe raster data}

\description{
Describe a raster data set with gdallinfo
}


\usage{
fwInfo(x, options)
}

\arguments{
  \item{x}{Character (filename), Raster* or SpatialGrid* object}
  \item{options}{options. See \url{http://www.gdal.org/gdalinfo.html} }  
}


\value{
character vector 
}

\seealso{
\url{http://www.gdal.org/gdalinfo.html} 
}


\author{Robert J. Hijmans}


\keyword{spatial}

